+++ CHANGES.rst
@@ -23,6 +23,10 @@ New features
 Major changes
 -------------
 
+* AggJoiner, AggTarget and MultiAggJoiner now require the operations argument. They do not
+  split columns by type anymore, but apply operations on all selected cols. :pr:`1116`
+  by :user:`Théo Jolivet <TheooJ>`.
+
 Minor changes
 -------------